backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.9
python-multipart==0.0.12
pillow==10.4.0
pypdf==4.3.1
python-docx==1.1.2
pytesseract==0.3.13
chromadb==0.5.18
sentence-transformers==3.1.1
ollama==0.3.3
python-magic==0.4.27
exifread==3.0.0
insightface==0.7.3
onnxruntime==1.19.2
python-dotenv==1.0.1
aiofiles==24.1.0
watchfiles==0.24.0
